Moves getAmpAdResourceId() from service.js to ad-helper.js
